Add ability to bath load data based on pagination, so when have 100 items it will load 25 items at a time or pageSize provided. Also took opportunity to refactor the code to make it easier to manage
