+++ src/utils/generate-zod-schema.ts
@@ -1,0 +1,79 @@
+import { z } from "zod";
+import type { ZodTypeAny, ZodObject } from "zod";
+
+
+/**
+ * Generates a Zod schema from Strapi content type attributes
+ * @param attributes The Strapi content type attributes
+ * @returns A Zod object schema representing the content type
+ */
+
+function generateZodSchema(attributes: Record<string, any>): ZodObject<any> {
+  const shape: Record<string, ZodTypeAny> = {};
+  for (const [key, value] of Object.entries(attributes)) {
+    const { type, ...rest } = value;
+    shape[key] = mapTypeToZodSchema(type, rest);
+  }
+  return z.object(shape);
+}
+
+
+/**
+ * Maps Strapi field types to Zod schema types
+ * @param type The Strapi field type
+ * @param field The field configuration object
+ * @returns A Zod schema corresponding to the Strapi field type
+ */
+
+function mapTypeToZodSchema(type: string, field: any): ZodTypeAny {
+  const schemaMap: Record<string, () => ZodTypeAny> = {
+    string: () => z.string(),
+    uid: () => z.string(),
+    media: () =>
+      z.object({
+        allowedTypes: z.array(z.enum(field.allowedTypes)),
+        type: z.literal("media"),
+        multiple: z.boolean(),
+        url: z.string(),
+        alternativeText: z.string().optional(),
+        caption: z.string().optional(),
+        width: z.number().optional(),
+        height: z.number().optional(),
+      }),
+    richtext: () => z.string(),
+    datetime: () => z.string().datetime(),
+    relation: () =>
+      z
+        .object({
+          relation: z.literal(field.relation),
+          target: z.literal(field.target),
+          configurable: z.boolean().optional(),
+          writable: z.boolean().optional(),
+          visible: z.boolean().optional(),
+          useJoinTable: z.boolean().optional(),
+          private: z.boolean().optional(),
+        })
+        .optional(),
+    boolean: () => z.boolean(),
+    number: () => z.number(),
+    array: () => z.array(mapTypeToZodSchema(field.items.type, field.items)),
+    object: () => {
+      const shape: Record<string, ZodTypeAny> = {};
+      for (const [key, value] of Object.entries(field.properties)) {
+        if (typeof value === "object" && value !== null && "type" in value) {
+          shape[key] = mapTypeToZodSchema(value.type as string, value);
+        } else {
+          throw new Error(`Invalid field value for key: ${key}`);
+        }
+      }
+      return z.object(shape);
+    },
+    text: () => z.string(),
+    dynamiczone: () => z.array(z.object({ __component: z.string() })),
+  };
+
+  return (schemaMap[type] || (() => z.any()))();
+}
+
+
+export { generateZodSchema };
